feat(GOV,TE): Phase 6 prep - verification artifacts gate, T3 deferred record, soak driver
- check_verification_artifacts.py: redesigned to pass when at least one
  artifact matches current HEAD; historical artifacts from prior HEADs
  are normal record-keeping, not stale
- docs/delivery/2026-04-27-bbc0c47-rule15-volces.json: T3 deferred record
  at bbc0c47; status=deferred, cap_factor=t3_deferred, cap_value=72.0
- docs/verification/bbc0c47-default-offline-clean-env.json: clean-env
  evidence at HEAD (8596 passed, default-offline profile)
- docs/verification/bbc0c47-observability-spine.json: spine evidence
- docs/verification/6ac41b4-observability-spine.json: prior-wave evidence
- docs/verification/6e11eea-chaos-matrix.json,
  docs/verification/e5b5183-chaos-matrix.json: chaos evidence
- scripts/soak_24h.py: III-2 rewrite - time-budget loop, _SamplerThread,
  real recovery counters, sigterm injection, --run-interval-seconds flag;
  RUF046 fix (round() returns int directly)
- tests/unit/scripts/test_soak_driver_loop.py: 3 tests; F401 fix (unused
  pytest import removed)

Owner: GOV,TE

+++ scripts/soak_24h.py
@@ -221,7 +221,7 @@ def _write_evidence(
     )
 
     label = "soak-dry" if dry_run else "soak"
-    duration_min = int(round(duration_seconds / 60))
+    duration_min = round(duration_seconds / 60)
     filename = f"{sha}-{label}-{duration_min}m.json"
     out_path = out_dir / filename
